+++ src/core/emsesp_common.h
