+++ src/core/helpers.h
@@ -20,7 +20,7 @@
 #define EMSESP_HELPERS_H
 
 #include "telegram.h" // for EMS_VALUE_* settings
-#include "common.h"
+#include "emsesp_common.h"
 
 namespace emsesp {
